Support header length byte in ledger pubkey request & response
The ledger app expects a “length” byte for both request and response.
This seems a little redundant, since the APDU itself has its own
length header. But, it’s easier to fix it here than to fix it on the
Ledger. Plus we get some reuse in data types when we embed addresses
in transactions.